Picking + camera control + related changes

+++ src/frame/input.rs
@@ -22,7 +22,7 @@ pub struct FrameInput {
     pub window_height: usize,
 
     /// Number of physical pixels for each logical pixel.
-    pub device_pixel_ratio: usize,
+    pub device_pixel_ratio: f64,
 
     /// Whether or not this is the first frame.
     pub first_frame: bool,
